# [js/webgpu] Add input/output shapes information to profiling