@fs-eire 2023-08-30
can we put tensor data type as well? they are in TensorView.dataType. It's a number and can be used with tensorDataTypeEnumToString() to convert to string representation.