@qjia7 2023-08-30
> can we put tensor data type as well? they are in TensorView.dataType. It's a number and can be used with tensorDataTypeEnumToString() to convert to string representation.

Done.
```
[profiling] kernel "[Split] 51288384" input[0]: 1,256,64,64 | float32, output[0]: 1,256,64,64 | float32, execution time: 31406 ns
program-manager.ts:115 
[profiling] kernel "[Concat] 52361040" input[0]: 1,256,64,64 | float32, output[0]: 1,256,64,64 | float32, execution time: 42968 ns
program-manager.ts:115 
[profiling] kernel "[Transpose] 52375264" input[0]: 1,256,64,64 | float32, output[0]: 1,64,64,256 | float32, execution time: 84167 ns
program-manager.ts:115 
[profiling] kernel "[Sub] 51098472" input[0]:  | float32, input[1]: 1 | float32, output[0]: 1 | float32, execution time: 6511 ns
program-manager.ts:115 
[profiling] kernel "[Mul] 51344440" input[0]: 1 | float32, input[1]: 1,256,1,1 | float32, output[0]: 1,256,1,1 | float32, execution time: 6771 ns