@qjia7 2023-08-30
I just notices that @guschmue 's profiling result is like 
```
[profiling] kernel "44184248|[Split] /Split_1" execution time: 1436672 ns
```
But my original result is like
```
[profiling] kernel "52097928|[Split] 52097928" execution time: 37760 ns
```
Not sure why the op name becomes kernel id. Anyway, I just keep the original profiling info unchanged and add the additional input shapes and output shapes.